moved firestore functions into a separate file
